emerging_infra_scraper/requirements.txt
beautifulsoup4==4.14.2
feedparser==6.0.12
PyYAML==6.0.3
Requests==2.32.5
